Move core interceptors into top level classes.
No behavior changes here.
