+++ okhttp/src/main/java/okhttp3/internal/http/HttpEngine.java
@@ -17,32 +17,22 @@
 package okhttp3.internal.http;
 
 import java.io.IOException;
-import java.io.InterruptedIOException;
-import java.net.ProtocolException;
-import java.net.Proxy;
-import java.net.SocketTimeoutException;
-import java.security.cert.CertificateException;
 import java.util.Date;
 import java.util.List;
 import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.SSLHandshakeException;
-import javax.net.ssl.SSLPeerUnverifiedException;
 import javax.net.ssl.SSLSocketFactory;
 import okhttp3.Address;
 import okhttp3.CertificatePinner;
-import okhttp3.Connection;
 import okhttp3.Cookie;
 import okhttp3.CookieJar;
 import okhttp3.Headers;
 import okhttp3.HttpUrl;
-import okhttp3.Interceptor;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
 import okhttp3.Protocol;
 import okhttp3.Request;
 import okhttp3.Response;
 import okhttp3.ResponseBody;
-import okhttp3.Route;
 import okhttp3.internal.Internal;
 import okhttp3.internal.InternalCache;
 import okhttp3.internal.Version;
@@ -55,22 +45,13 @@ import okio.Sink;
 import okio.Source;
 import okio.Timeout;
 
-import static java.net.HttpURLConnection.HTTP_CLIENT_TIMEOUT;
-import static java.net.HttpURLConnection.HTTP_MOVED_PERM;
-import static java.net.HttpURLConnection.HTTP_MOVED_TEMP;
-import static java.net.HttpURLConnection.HTTP_MULT_CHOICE;
 import static java.net.HttpURLConnection.HTTP_NOT_MODIFIED;
 import static java.net.HttpURLConnection.HTTP_NO_CONTENT;
-import static java.net.HttpURLConnection.HTTP_PROXY_AUTH;
-import static java.net.HttpURLConnection.HTTP_SEE_OTHER;
-import static java.net.HttpURLConnection.HTTP_UNAUTHORIZED;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static okhttp3.internal.Util.closeQuietly;
 import static okhttp3.internal.Util.discard;
 import static okhttp3.internal.Util.hostHeader;
 import static okhttp3.internal.http.StatusLine.HTTP_CONTINUE;
-import static okhttp3.internal.http.StatusLine.HTTP_PERM_REDIRECT;
-import static okhttp3.internal.http.StatusLine.HTTP_TEMP_REDIRECT;
 
 /**
  * Handles a single HTTP request/response pair. The request and response may be served by the HTTP
@@ -98,10 +79,8 @@ public final class HttpEngine {
   };
 
   final OkHttpClient client;
-
-  public final StreamAllocation streamAllocation;
-  private final Response priorResponse;
-  private HttpStream httpStream;
+  final StreamAllocation streamAllocation;
+  final Response priorResponse;
 
   /**
    * True if this client added an "Accept-Encoding: gzip" header field and is therefore responsible
@@ -110,15 +89,12 @@ public final class HttpEngine {
   private boolean transparentGzip;
 
   /** The original application-provided request URL. Never modified by OkHttp. */
-  private final HttpUrl userRequestUrl;
+  final HttpUrl userRequestUrl;
 
-  private final boolean forWebSocket;
-
-  public HttpEngine(OkHttpClient client, HttpUrl userRequestUrl,
-      boolean forWebSocket, StreamAllocation streamAllocation, Response priorResponse) {
+  public HttpEngine(OkHttpClient client, HttpUrl userRequestUrl, StreamAllocation streamAllocation,
+      Response priorResponse) {
     this.client = client;
     this.userRequestUrl = userRequestUrl;
-    this.forWebSocket = forWebSocket;
     this.streamAllocation = streamAllocation != null
         ? streamAllocation
         : new StreamAllocation(client.connectionPool(), createAddress(client, userRequestUrl));
@@ -130,13 +106,10 @@ public final class HttpEngine {
    * Prepares the request headers and gets ready to start writing the request body if it exists.
    *
    * @throws RouteException if the was a problem during connection via a specific route. Sometimes
-   * recoverable. See {@link #recover}.
-   * @throws IOException if there was a problem while making a request. Sometimes recoverable. See
-   * {@link #recover}.
+   *     recoverable.
+   * @throws IOException if there was a problem while making a request. Sometimes recoverable.
    */
   public Response proceed(Request userRequest, RealInterceptorChain chain) throws IOException {
-    if (httpStream != null) throw new IllegalStateException();
-
     Request request = networkRequest(userRequest);
 
     InternalCache responseCache = Internal.instance.internalCache(client);
@@ -183,14 +156,13 @@ public final class HttpEngine {
     }
 
     // We need the network to satisfy this request. Possibly for validating a conditional GET.
-    boolean success = false;
+    HttpStream httpStream = null;
     try {
       httpStream = connect(networkRequest);
       httpStream.setHttpEngine(this);
-      success = true;
     } finally {
       // If we're crashing on I/O or otherwise, don't leak the cache body.
-      if (!success && cacheCandidate != null) {
+      if (httpStream == null && cacheCandidate != null) {
         closeQuietly(cacheCandidate.body());
       }
     }
@@ -211,7 +183,7 @@ public final class HttpEngine {
             .networkResponse(stripBody(networkResponse))
             .build();
         networkResponse.body().close();
-        releaseStreamAllocation();
+        streamAllocation.release();
 
         // Update the cache after combining headers but before stripping the
         // Content-Encoding header (as performed by initContentStream()).
@@ -252,72 +224,6 @@ public final class HttpEngine {
         : response;
   }
 
-  public Connection getConnection() {
-    return streamAllocation.connection();
-  }
-
-  /**
-   * Report and attempt to recover from a failure to communicate with a server. Returns a new HTTP
-   * engine that should be used for the retry if {@code e} is recoverable, or null if the failure is
-   * permanent. Requests with a body can only be recovered if the body is buffered.
-   */
-  public HttpEngine recover(IOException e, boolean routeException, Request userRequest) {
-    streamAllocation.streamFailed(e);
-
-    if (!client.retryOnConnectionFailure()) {
-      return null; // The application layer has forbidden retries.
-    }
-
-    if (!routeException && userRequest.body() instanceof UnrepeatableRequestBody) {
-      return null; // We can't send the request body again.
-    }
-
-    if (!isRecoverable(e, routeException)) {
-      return null; // This exception is fatal.
-    }
-
-    if (!streamAllocation.hasMoreRoutes()) {
-      return null; // No more routes to attempt.
-    }
-
-    StreamAllocation streamAllocation = close(null);
-
-    // For failure recovery, use the same route selector with a new connection.
-    return new HttpEngine(client, userRequestUrl, forWebSocket, streamAllocation, priorResponse);
-  }
-
-  private boolean isRecoverable(IOException e, boolean routeException) {
-    // If there was a protocol problem, don't recover.
-    if (e instanceof ProtocolException) {
-      return false;
-    }
-
-    // If there was an interruption don't recover, but if there was a timeout connecting to a route
-    // we should try the next route (if there is one).
-    if (e instanceof InterruptedIOException) {
-      return e instanceof SocketTimeoutException && routeException;
-    }
-
-    // Look for known client-side or negotiation errors that are unlikely to be fixed by trying
-    // again with a different route.
-    if (e instanceof SSLHandshakeException) {
-      // If the problem was a CertificateException from the X509TrustManager,
-      // do not retry.
-      if (e.getCause() instanceof CertificateException) {
-        return false;
-      }
-    }
-    if (e instanceof SSLPeerUnverifiedException) {
-      // e.g. a certificate pinning error.
-      return false;
-    }
-
-    // An example of one we might want to retry with a different route is a problem connecting to a
-    // proxy and would manifest as a standard IOException. Unless it is one we know we should not
-    // retry, we return true and try a new route.
-    return true;
-  }
-
   private CacheRequest maybeCache(Response userResponse, Request networkRequest,
       InternalCache responseCache) throws IOException {
     if (responseCache == null) return null;
@@ -336,27 +242,6 @@ public final class HttpEngine {
 
     // Offer this request to the cache.
     return responseCache.put(userResponse);
-  }
-
-  /**
-   * Configure the socket connection to be either pooled or closed when it is either exhausted or
-   * closed. If it is unneeded when this is called, it will be released immediately.
-   */
-  public void releaseStreamAllocation() throws IOException {
-    streamAllocation.release();
-  }
-
-  /**
-   * Immediately closes the socket connection if it's currently held by this engine. Use this to
-   * interrupt an in-flight request from any thread. It's the caller's responsibility to close the
-   * request body and response body streams; otherwise resources may be leaked.
-   *
-   * <p>This method is safe to be called concurrently, but provides limited guarantees. If a
-   * transport layer connection has been established (such as a HTTP/2 stream) that is terminated.
-   * Otherwise if a socket connection is being established, that is terminated.
-   */
-  public void cancel() {
-    streamAllocation.cancel();
   }
 
   /**
@@ -387,7 +272,7 @@ public final class HttpEngine {
    * Modified" can include "Content-Encoding: gzip" without a response body and we will crash if we
    * attempt to decompress the zero-byte source.
    */
-  private Response unzip(final Response response) throws IOException {
+  private Response unzip(Response response) throws IOException {
     if (!transparentGzip || !"gzip".equalsIgnoreCase(response.header("Content-Encoding"))) {
       return response;
     }
@@ -479,68 +364,6 @@ public final class HttpEngine {
       cookieHeader.append(cookie.name()).append('=').append(cookie.value());
     }
     return cookieHeader.toString();
-  }
-
-  public static class CallServerInterceptor implements Interceptor {
-    private final boolean forWebSocket;
-
-    /** The time when the request headers were written, or -1 if they haven't been written yet. */
-    long sentRequestMillis = -1;
-
-    public CallServerInterceptor(boolean forWebSocket) {
-      this.forWebSocket = forWebSocket;
-    }
-
-    @Override public Response intercept(Chain chain) throws IOException {
-      HttpStream httpStream = ((RealInterceptorChain) chain).httpStream();
-      StreamAllocation streamAllocation = ((RealInterceptorChain) chain).streamAllocation();
-
-      Request request = chain.request();
-      sentRequestMillis = System.currentTimeMillis();
-      httpStream.writeRequestHeaders(request);
-
-      if (HttpMethod.permitsRequestBody(request.method()) && request.body() != null) {
-        Sink requestBodyOut = httpStream.createRequestBody(request, request.body().contentLength());
-        BufferedSink bufferedRequestBody = Okio.buffer(requestBodyOut);
-        request.body().writeTo(bufferedRequestBody);
-        bufferedRequestBody.close();
-      }
-
-      Response response = readNetworkResponse(httpStream, streamAllocation, request);
-
-      int code = response.code();
-      if ((code == 204 || code == 205) && response.body().contentLength() > 0) {
-        throw new ProtocolException(
-            "HTTP " + code + " had non-zero Content-Length: " + response.body().contentLength());
-      }
-
-      return response;
-    }
-
-    private Response readNetworkResponse(HttpStream httpStream, StreamAllocation streamAllocation,
-        Request networkRequest) throws IOException {
-      httpStream.finishRequest();
-
-      Response networkResponse = httpStream.readResponseHeaders()
-          .request(networkRequest)
-          .handshake(streamAllocation.connection().handshake())
-          .sentRequestAtMillis(sentRequestMillis)
-          .receivedResponseAtMillis(System.currentTimeMillis())
-          .build();
-
-      if (!forWebSocket || networkResponse.code() != 101) {
-        networkResponse = networkResponse.newBuilder()
-            .body(httpStream.openResponseBody(networkResponse))
-            .build();
-      }
-
-      if ("close".equalsIgnoreCase(networkResponse.request().header("Connection"))
-          || "close".equalsIgnoreCase(networkResponse.header("Connection"))) {
-        streamAllocation.noNewStreams();
-      }
-
-      return networkResponse;
-    }
   }
 
   /**
@@ -666,107 +489,6 @@ public final class HttpEngine {
     if (cookies.isEmpty()) return;
 
     client.cookieJar().saveFromResponse(userRequestUrl, cookies);
-  }
-
-  /**
-   * Figures out the HTTP request to make in response to receiving this engine's response. This will
-   * either add authentication headers, follow redirects or handle a client request timeout. If a
-   * follow-up is either unnecessary or not applicable, this returns null.
-   */
-  public Request followUpRequest(Response userResponse) throws IOException {
-    if (userResponse == null) throw new IllegalStateException();
-    Connection connection = streamAllocation.connection();
-    Route route = connection != null
-        ? connection.route()
-        : null;
-    int responseCode = userResponse.code();
-
-    final String method = userResponse.request().method();
-    switch (responseCode) {
-      case HTTP_PROXY_AUTH:
-        Proxy selectedProxy = route != null
-            ? route.proxy()
-            : client.proxy();
-        if (selectedProxy.type() != Proxy.Type.HTTP) {
-          throw new ProtocolException("Received HTTP_PROXY_AUTH (407) code while not using proxy");
-        }
-        return client.proxyAuthenticator().authenticate(route, userResponse);
-
-      case HTTP_UNAUTHORIZED:
-        return client.authenticator().authenticate(route, userResponse);
-
-      case HTTP_PERM_REDIRECT:
-      case HTTP_TEMP_REDIRECT:
-        // "If the 307 or 308 status code is received in response to a request other than GET
-        // or HEAD, the user agent MUST NOT automatically redirect the request"
-        if (!method.equals("GET") && !method.equals("HEAD")) {
-          return null;
-        }
-        // fall-through
-      case HTTP_MULT_CHOICE:
-      case HTTP_MOVED_PERM:
-      case HTTP_MOVED_TEMP:
-      case HTTP_SEE_OTHER:
-        // Does the client allow redirects?
-        if (!client.followRedirects()) return null;
-
-        String location = userResponse.header("Location");
-        if (location == null) return null;
-        HttpUrl url = userResponse.request().url().resolve(location);
-
-        // Don't follow redirects to unsupported protocols.
-        if (url == null) return null;
-
-        // If configured, don't follow redirects between SSL and non-SSL.
-        boolean sameScheme = url.scheme().equals(userResponse.request().url().scheme());
-        if (!sameScheme && !client.followSslRedirects()) return null;
-
-        // Redirects don't include a request body.
-        Request.Builder requestBuilder = userResponse.request().newBuilder();
-        if (HttpMethod.permitsRequestBody(method)) {
-          if (HttpMethod.redirectsToGet(method)) {
-            requestBuilder.method("GET", null);
-          } else {
-            requestBuilder.method(method, null);
-          }
-          requestBuilder.removeHeader("Transfer-Encoding");
-          requestBuilder.removeHeader("Content-Length");
-          requestBuilder.removeHeader("Content-Type");
-        }
-
-        // When redirecting across hosts, drop all authentication headers. This
-        // is potentially annoying to the application layer since they have no
-        // way to retain them.
-        if (!sameConnection(userResponse, url)) {
-          requestBuilder.removeHeader("Authorization");
-        }
-
-        return requestBuilder.url(url).build();
-
-      case HTTP_CLIENT_TIMEOUT:
-        // 408's are rare in practice, but some servers like HAProxy use this response code. The
-        // spec says that we may repeat the request without modifications. Modern browsers also
-        // repeat the request (even non-idempotent ones.)
-        if (userResponse.request().body() instanceof UnrepeatableRequestBody) {
-          return null;
-        }
-
-        return userResponse.request();
-
-      default:
-        return null;
-    }
-  }
-
-  /**
-   * Returns true if an HTTP request for {@code followUp} can reuse the connection used by this
-   * engine.
-   */
-  public boolean sameConnection(Response response, HttpUrl followUp) {
-    HttpUrl url = response.request().url();
-    return url.host().equals(followUp.host())
-        && url.port() == followUp.port()
-        && url.scheme().equals(followUp.scheme());
   }
 
   private static Address createAddress(OkHttpClient client, HttpUrl url) {
